+++ test/cadet_web/admin_controllers/admin_goals_controller_test.exs
@@ -263,6 +263,6 @@ defmodule CadetWeb.AdminGoalsControllerTest do
   end
 
   defp build_path(course_id, uuid, course_reg_id) do
-    "v2/courses/#{course_id}/admin/users/#{course_reg_id}/goals/#{uuid}/progress/"
+    "/v2/courses/#{course_id}/admin/users/#{course_reg_id}/goals/#{uuid}/progress/"
   end
 end
